refactor undiscounted price to use the same mechanism
